Rename manual_testing to file_io
